Conditional Independence Test

We test whether x and y are conditionally associated, given S, using specific methods.

ortest(x, y, S, suffStat)

Arguments

x

The position of the exposure variable

y

The position of the exposure variable

S

The position of the conditional variable set

suffStat

A list with elements needed for the test:

"dat"

Data matrix where columns represent variables.

"method"

The method used for the test: "sl", "linear", "linear_int"

"sl"

Specifies the super learning model. If y is numeric, the default method is linear regression model, mean response,MARS, random forest and XGBoost. If y is binary, the default method is LDA, mean response,MARS, random forest and XGBoost.

"cross_fitting"

Logical; whether cross-fitting is used in super learning.

"kfolds"

Number of folds for cross-fitting.

"two_way"

Logical; whether to include two-way interactions between variables in S.

"three_way"

Logical; whether to include three-way interactions between variables in S.

Value

p-value for model y ~ x + S.

Details

This function test whether x and y is independent conditional on S. The final result only includes p value for model y ~ x + S.